pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cheflink"
version = "0.1.0"
description = "ChefLink"
authors = [{name = "Yuhshun Chen", email = "yuhshun@chen.zone"}]
dependencies = [
    "fastapi",
    "sqlalchemy",
    "python-dotenv",
    "psycopg2-binary",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"